requirements.txt
flask
torch
torchvision
numpy
pillow
typing
requests
matplotlib
docker
